make Maxim's panning algorithms use 16.16 fixed point
This is more or less a cosmetic change, but in general I prefer fixed point
calculations over floating point.
(It also eliminates int->float->int conversions.)
Also made the PI/SQRT2 constants slightly more accurate.

+++ emu/cores/emu2413.h
@@ -75,7 +75,7 @@ typedef struct __OPLL {
   uint32_t opllstep ;
   //int32_t prev, next ;
   int32_t sprev[2],snext[2];
-  float pan[15][2];
+  int32_t pan[15][2];
 
   /* Register */
   uint8_t reg[0x40] ;
